frameworks/python/django/12_testblog/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
markdown==2.6.6 
django-filter==2.1.0
psycopg2-binary==2.8.1
coreapi==2.3.3
pytz==2018.9